Implement initial database unit tests
This is just a basic test for schema migration, and a basic test for the
enlarger profile DAO. These are not meant to be extensive, or even
useful. Rather, they are to put the framework in place so that its
easier to write good tests when needed in the future.