Initial implementation of paper profile management
Added database code to handle storage of paper profiles

Added database(ish) code to wrap a local JSON asset with a collection of
stock paper profiles

Built out the UI components for managing the list of profiles,
adding/editing items on the list of profiles, and adding to the profile
list from the stock data.

Still need to implement more input validation, and round out all the
corners of the profile management use cases.