Stock profile data for Ilford papers
This is a first pass at baking in data for commonly used papers, and
includes content from the official datasheets.
While most of these papers are multigrade, there are a few fixed grade
ones in there. The fixed-grade papers will look the same, but may omit
data for grades that aren't available.