Paper grade recommendations and paper settings dialog data rewrite
Changed all paper-related home fragment dialogs to use a
PaperProfileEntity object (now Parcelable) directly, rather than having
to rely on repeatedly reloading async objects from the DB layer.

This allows the code inside these components to be simplified
significantly, and avoids the potential for a lot of fiddly data
interaction bugs that are difficult to squash.

The dialog for the larger paper profile now also takes a "Reference
ISO(R)" value from the smaller paper profile, which allows it to
automatically select a starting grade when the user changes papers.