Drop print size relationship requirements
There is no need for the "target" print to actually at an equal or
larger height than the "base" print. We can even adjust from a big print
to a small print. So now that the terminology in the app is changed, the
validation code should also be changed.

Fixes #4