Change mode-switching recalculation behavior on burn/dodge dialog
Track the last adjustment mode that actually had user-provided input,
and use that as a basis for adjustment recalculation when switching
input modes.

This makes it so, for example, if a user enters a percentage, then
switches to stops, then switches back to percent, their percent value
doesn't change. (Normally it would, because the stops conversion has to
round its result to constrain it to adjustment increments, so converting
back would be lossy.